requirements.txt
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.2.0
googlesearch-python==1.2.3
idna==3.4
requests==2.31.0
soupsieve==2.4.1
urllib3==2.0.3